requirements.txt
argon2==0.1.10
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
bcrypt==4.1.2
black==24.1.1
cffi==1.16.0
click==8.1.7
dj-database-url==2.1.0
Django==4.1.4
django-address==0.2.8
django-on-heroku==1.1.2
Faker==22.6.0
gunicorn==21.2.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
whitenoise==6.6.0
